Move implementation into rktest.h behind ifdef
